Fixed and changed some stuff (desc)
Fixed ghost player issue on rejoining players,
Fixed the animator on the enemies,
Added blending on the attack animation on the enemy,
Player no longer regenerates hp when moving,
Player no longer regenerates hp or mana while in the upgrade menu.
Fixed the Gameover screen for a simpler and more failsafe fix.